- disable `ovelay limit` hooks , still crashing game - sync to latest develop - Fix crash when selecting unit/infantry with `HasRadialIndicator` - Fix missing code on SW drain click code - Fix crash caused by broken aplha code - Fix bug where techno health set to `INT_MIN` when hit by radioation with 0 distance and 0 spread - Fix `WallTowers` ini reading is missing - Fix crash when `CreateUnit` spawning unit on top of bridge - Fix bug where techno only able to damage once and never get damaged again due to broken code - Fix tiberium pip color is flipped - Fix crash caused by dangling pointer of `BulletExt` `AttachedSystem` pointer - Fix `GenericPrerequisite` not save/loaded properly - Fix `HelperedVector` not save/loaded properly - revert Fix `EMPulseCannon=yes` building because it causing building uneable to facing target when firing `EMPulse` SW - Fix anim ownership problem for some mod - change some code style - add `WindDirectionMult` customization - add `VeinsDamagingWeightTreshold` - add `VeinholePalette` - add `Weeder.TriggerPreProductionBuildingAnim` - add `Weeder.PipIndex` - add `Weeder.PipEmptyIndex` - add option to disable `GenerateColorSpread_ShadeCountSet` fix , it breaking animations on certain mod

+++ src/Misc/Ares/Hooks/Hooks.Verses.cpp
@@ -57,7 +57,7 @@ DEFINE_OVERRIDE_HOOK(0x489235, GetTotalDamage_Verses, 0x8)
 	const auto pExt = WarheadTypeExt::ExtMap.Find(pWH);
 
 
-	if (nArmor > ArmorTypeClass::Array.size())
+	if ((size_t)nArmor > ArmorTypeClass::Array.size())
 		Debug::Log(__FUNCTION__" Armor is more that avaible ArmorTypeClass \n");
 
 	const auto vsData = &pExt->Verses[nArmor];
@@ -76,7 +76,7 @@ DEFINE_OVERRIDE_HOOK(0x6F7D3D, TechnoClass_CanAutoTargetObject_Verses, 0x7)
 
 	const auto pData = WarheadTypeExt::ExtMap.Find(pWH);
 
-	if (nArmor > ArmorTypeClass::Array.size())
+	if ((size_t)nArmor > ArmorTypeClass::Array.size())
 		Debug::Log(__FUNCTION__" Armor is more that avaible ArmorTypeClass \n");
 
 	const auto vsData = &pData->Verses[nArmor];
@@ -98,7 +98,7 @@ DEFINE_OVERRIDE_HOOK(0x6FCB6A, TechnoClass_CanFire_Verses, 0x7)
 	const auto pData = WarheadTypeExt::ExtMap.Find(pWH);
 	const auto vsData = &pData->Verses[nArmor];
 
-	if (nArmor > ArmorTypeClass::Array.size())
+	if ((size_t)nArmor > ArmorTypeClass::Array.size())
 		Debug::Log(__FUNCTION__" Armor is more that avaible ArmorTypeClass \n");
 
 	// i think there is no way for the techno know if it attack using force fire or not
@@ -157,7 +157,7 @@ DEFINE_OVERRIDE_HOOK(0x70CF45, TechnoClass_EvalThreatRating_ThisWeaponWarhead_Ve
 
 	const auto pData = WarheadTypeExt::ExtMap.Find(pWH);
 
-	if (nArmor > ArmorTypeClass::Array.size())
+	if ((size_t)nArmor > ArmorTypeClass::Array.size())
 		Debug::Log(__FUNCTION__" Armor is more that avaible ArmorTypeClass \n");
 
 	const auto vsData = &pData->Verses[nArmor];
@@ -180,7 +180,7 @@ DEFINE_OVERRIDE_HOOK(0x6F36E3, TechnoClass_SelectWeapon_Verses, 0x5)
 	GET_STACK(WeaponTypeClass*, pPrimary, 0x14); //primary
 
 	const int nArmor = (int)pTarget->GetTechnoType()->Armor;
-	if (nArmor > ArmorTypeClass::Array.size())
+	if ((size_t)nArmor > ArmorTypeClass::Array.size())
 		Debug::Log(__FUNCTION__" Armor is more that avaible ArmorTypeClass \n");
 
 	const auto vsData_Secondary = &WarheadTypeExt::ExtMap.Find(pSecondary->Warhead)->Verses[nArmor];
@@ -201,7 +201,7 @@ DEFINE_OVERRIDE_HOOK(0x708AF7, TechnoClass_ShouldRetaliate_Verses, 0x7)
 	GET(int, nArmor, EAX);
 
 	const auto pData = WarheadTypeExt::ExtMap.Find(pWH);
-	if (nArmor > ArmorTypeClass::Array.size())
+	if ((size_t)nArmor > ArmorTypeClass::Array.size())
 		Debug::Log(__FUNCTION__" Armor is more that avaible ArmorTypeClass \n");
 
 	const auto vsData = &pData->Verses[nArmor];
